stabilityai/requirements.txt
fastapi==0.95.0                     # FastAPI framework
uvicorn==0.22.0                     # ASGI server for serving FastAPI
pydantic==1.10.4                    # Data validation and settings management using Python type annotations
python-multipart==0.0.6             # For handling form data, especially file uploads
torch==1.13.1                       # PyTorch for model loading and inference
diffusers==0.18.0                    # Library for diffusion models (e.g., Stable Diffusion)
huggingface-hub==0.14.1             # For accessing Hugging Face model hub
pdfminer.six==20221105               # For extracting text from PDF files
python-pptx==0.6.21                 # For creating PowerPoint presentations
google-auth==2.16.0                  # Google API authentication
google-api-python-client==2.81.0     # Google API client for accessing various Google services
requests==2.28.1                     # For making HTTP requests